+++ R/issues_descriptions.R
@@ -8,6 +8,8 @@
 #'     <https://github.com/RConsortium/r-advisory-database>.
 #'   2. Licenses that cannot be verified as free and open-source.
 #'   3. The presence of a `"Remotes"` field.
+#'   4. A lower version number on R-multiverse than on CRAN,
+#'     if the package is also published there.
 #' @inheritSection record_issues Package issues
 #' @return A named list of information about packages which do not comply
 #'   with `DESCRPTION` checks. Each name is a package name,
@@ -24,8 +26,16 @@ issues_descriptions <- function(meta = meta_packages()) {
   meta <- issues_descriptions_advisories(meta)
   meta <- issues_descriptions_licenses(meta)
   meta <- issues_descriptions_remotes(meta)
+  meta <- issues_descriptions_version_conflict(meta, repo = "cran")
   meta <- meta[meta$issue,, drop = FALSE] # nolint
-  issues_list(meta[, c("package", "advisories", "license", "remotes")])
+  fields <- c(
+    "package",
+    "advisories",
+    "license",
+    "remotes",
+    "cran"
+  )
+  issues_list(meta[, fields])
 }
 
 issues_descriptions_advisories <- function(meta) {
@@ -51,5 +61,21 @@ issues_descriptions_remotes <- function(meta) {
   meta[["remotes"]] <- meta[["remotes"]] %||% replicate(nrow(meta), NULL)
   meta$remotes <- lapply(meta$remotes, function(x) x[nzchar(x)])
   meta$issue <- meta$issue | vapply(meta$remotes, length, integer(1L)) > 0L
+  meta
+}
+
+issues_descriptions_version_conflict <- function(meta, repo) {
+  conflict <- vapply(
+    X = seq_len(nrow(meta)),
+    FUN = function(index) {
+      utils::compareVersion(
+        a = meta$version[index],
+        b = meta[[repo]][index]
+      ) < 0L
+    },
+    FUN.VALUE = logical(1L)
+  )
+  meta[[repo]][!conflict] <- NA_character_
+  meta$issue <- meta$issue | conflict
   meta
 }