+++ R/meta_packages.R
@@ -3,7 +3,18 @@
 #' @family meta
 #' @description List package metadata in an R universe.
 #' @return A data frame with one row per package and columns with package
-#'   metadata.
+#'   metadata. The most important columns are:
+#'   * `package`: character vector of package names.
+#'   * `version`: character vector of package versions in the repo.
+#'   * `license`: character vector of license names.
+#'   * `remotesha`: character vector of GitHub/GitLab commit hashes.
+#'   * `remotes`: list of character vectors with dependencies in the
+#'     `Remotes:` field of the `DESCRIPTION` files.
+#'   * `foss`: `TRUE` if the package has a valid free open-source license,
+#'     `FALSE` otherwise.
+#'   * `cran`: character vector of versions. Each version is the version of
+#'     the package that was on CRAN during the first day of the most recent
+#'     R-multiverse staging period.
 #' @inheritParams meta_checks
 #' @param fields Character string of fields to query.
 #' @examples
@@ -29,5 +40,14 @@ meta_packages <- function(
   foss <- utils::available.packages(repos = repo, filters = "license/FOSS")
   out$foss <- FALSE
   out[as.character(foss[, "Package"]), "foss"] <- TRUE
+  freeze <- date_staging_freeze()
+  p3m <- "https://packagemanager.posit.co"
+  repo_cran <- file.path(p3m, "cran", freeze)
+  cran <- utils::available.packages(repos = repo_cran)
+  cran <- data.frame(
+    package = as.character(cran[, "Package"]),
+    cran = as.character(cran[, "Version"])
+  )
+  out <- merge(x = out, y = cran, all.x = TRUE, all.y = FALSE)
   out
 }