+++ tests/testthat/test-meta_packages.R
@@ -5,7 +5,8 @@ test_that("meta_packages()", {
   fields <- c(
     "package",
     "version",
-    "remotesha"
+    "remotesha",
+    "cran"
   )
   expect_true(all(fields %in% colnames(out)))
 })